Updated template input file; Linear iteration on single variable complete
Input file updated with stainless steel reflector region and correct boundary conditions.

Code now updates the reflector thickness with a linear ratio of the target k to the current k. This is handled in a single function so it could be modified in the future for more advanced methods (better extrapolation). Successive input files are created with updated iteration parameters with the intent of converging on the target criticality

Signed-off-by: drewejohnson <1drew.e.johnson@gmail.com>