Major change: Removed global parameters, Updated README and LICENSE
Removed the global parameter file. Everything is encompassed inside a dictionary that is passed between functions. The code as a whole seems more streamlined and connected (the output parser is contained in the iterator file). The code just feels better.

Did test the code and it appears to be working quite well. Will undergo more testing tomorrow. Tests should include some import tests.

Signed-off-by: drewejohnson <1drew.e.johnson@gmail.com>